re-arrange constants folder
